refactor: restructure files
