+++ libs/ngx-localstorage/src/index.ts
@@ -8,7 +8,7 @@ export * from './lib/services/ngx-localstorage.service';
 // directives
 export * from './lib/directives/ngx-localstorage.directive';
 // decorator
-export * from './lib/decorators';
+export * from './lib/decorators/ngx-localstorage.decorator';
 // interfaces
 export * from './lib/interfaces/decorator-options';
 export * from './lib/interfaces/storage-configuration';
