+++ libs/ngx-localstorage/src/lib/services/ngx-localstorage.service.ts
@@ -2,13 +2,14 @@ import { Injectable, Inject, OnDestroy, EventEmitter } from '@angular/core';
 
 import { NgxLocalstorageConfiguration } from '../interfaces/storage-configuration';
 import { PromisableService } from './promisable.service';
-import { constructKey, isSerializer } from '../utils';
 import { NGX_LOCAL_STORAGE_CONFIG } from '../tokens/storage-config';
 import { NGX_LOCAL_STORAGE_SERIALIZER } from '../tokens/storage-serializer';
 import { StorageSerializer } from '../interfaces/storage-serializer';
 import { filter, fromEvent, Observable, Subscription } from 'rxjs';
 import { STORAGE, STORAGE_SUPPORT } from '../tokens/storage';
 import { WINDOW } from '../tokens/window';
+import { isSerializer } from '../utils/guards';
+import { constructKey } from '../utils/key-utils';
 
 /**
  * Provides a service to access the localstorage.
